+++ modules/ui/fields/access.js
@@ -266,55 +266,67 @@ export function uiFieldAccess(field, context) {
         utilGetSetValue(items.selectAll('.preset-input-access'), function(d) {
                 return typeof tags[d] === 'string' ? tags[d] : '';
             })
-            .classed('mixed', function(d) {
-                return tags[d] && Array.isArray(tags[d]);
+            .classed('mixed', function(accessField) {
+                return tags[accessField] && Array.isArray(tags[accessField])
+                    || new Set(getAllPlaceholders(tags, accessField)).size > 1;
             })
-            .attr('title', function(d) {
-                return tags[d] && Array.isArray(tags[d]) && tags[d].filter(Boolean).join('\n');
+            .attr('title', function(accessField) {
+                return tags[accessField] && Array.isArray(tags[accessField]) && tags[accessField].filter(Boolean).join('\n');
             })
-            .attr('placeholder', function(d) {
-                if (tags[d] && Array.isArray(tags[d])) {
+            .attr('placeholder', function(accessField) {
+                let placeholders = getAllPlaceholders(tags, accessField);
+                if (new Set(placeholders).size === 1) {
+                    // all objects have the same implied access
+                    return placeholders[0];
+                } else {
                     return t('inspector.multiple_values');
                 }
-                if (d === 'bicycle' || d === 'motor_vehicle') {
-                    if (tags.vehicle && typeof tags.vehicle === 'string') {
-                        return tags.vehicle;
-                    }
+            });
+
+            function getAllPlaceholders(tags, accessField) {
+                let allTags = tags[Symbol.for('allTags')];
+                if (allTags && allTags.length > 1) {
+                    // multi selection
+                    const placeholders = [];
+                    allTags.forEach(tags => {
+                        placeholders.push(getPlaceholder(tags, accessField));
+                    });
+                    return placeholders;
+                } else {
+                    return [getPlaceholder(tags, accessField)];
                 }
-                if (tags.access && typeof tags.access === 'string') {
+            }
+
+            function getPlaceholder(tags, accessField) {
+                if (tags[accessField]) {
+                    return tags[accessField];
+                }
+                // implied access
+                // motorroad: https://wiki.openstreetmap.org/wiki/OSM_tags_for_routing/Access_restrictions
+                if (tags.motorroad === 'yes' && (accessField === 'foot' || accessField === 'bicycle' || accessField === 'horse')) {
+                    return 'no';
+                }
+                // inherited access
+                if (tags.vehicle && (accessField === 'bicycle' || accessField === 'motor_vehicle')) {
+                    return tags.vehicle;
+                }
+                if (tags.access) {
                     return tags.access;
                 }
-                function getPlaceholdersByTag(key, placeholdersByKey) {
-                    if (typeof tags[key] === 'string') {
-                        if (placeholdersByKey[tags[key]] &&
-                            placeholdersByKey[tags[key]][d]) {
-                            return placeholdersByKey[tags[key]][d];
-                        }
-                    } else {
-                        var impliedAccesses = tags[key].filter(Boolean).map(function(val) {
-                            return placeholdersByKey[val] && placeholdersByKey[val][d];
-                        }).filter(Boolean);
-
-                        if (impliedAccesses.length === tags[key].length &&
-                            new Set(impliedAccesses).size === 1) {
-                            // if all the barrier values have the same implied access for this type then use that
-                            return impliedAccesses[0];
-                        }
-                    }
-                }
+                // default access by road/barrier type
                 for (const key in placeholdersByTag) {
                     if (tags[key]) {
-                        const impliedAccess = getPlaceholdersByTag(key, placeholdersByTag[key]);
-                        if (impliedAccess) {
-                            return impliedAccess;
+                        if (placeholdersByTag[key][tags[key]] &&
+                            placeholdersByTag[key][tags[key]][accessField]) {
+                            return placeholdersByTag[key][tags[key]][accessField];
                         }
                     }
                 }
-                if (d === 'access' && !tags.barrier) {
+                if (accessField === 'access' && !tags.barrier) {
                     return 'yes';
                 }
                 return field.placeholder();
-            });
+            }
     };